+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds `labelAsAriaLabel` prop to `Menu` to fix incorrect ARIA label on button.
 - Adds `responsiveSpacing` object to `globals` to be used for implementing responsive spacing styles within DS components.
 
 ### Fixes